Moves nominatim api to its own file (from utils.js)
Move owner-api and nominatim-api to api subfolder
